+++ integration_tests/src/tests/mod.rs
@@ -10,7 +10,8 @@ mod change_owner_test;
 pub use change_owner_test::*;
 mod commit_post_test;
 pub use commit_post_test::*;
-// mod datacap_tests;
+mod datacap_tests;
+pub use datacap_tests::*;
 // mod evm_test;
 // mod extend_sectors_test;
 // mod init_test;
